Updating llm module to seperate concerns and keep modular
