+++ backend/src/knowledge_table_api/api/v1/endpoints/graph.py
@@ -14,7 +14,7 @@ from whyhow import Chunk, ChunkMetadata
 from knowledge_table_api.core.dependencies import get_llm_service
 from knowledge_table_api.models.graph import Table
 from knowledge_table_api.services.graph import generate_triples
-from knowledge_table_api.services.llm import generate_schema
+from knowledge_table_api.services.llm_operations import generate_schema
 
 router = APIRouter(tags=["Graph"], prefix="/graph")
