+++ backend/src/knowledge_table_api/services/graph.py
@@ -8,7 +8,7 @@ from whyhow import Node, Relation, Triple
 
 from knowledge_table_api.core.dependencies import get_llm_service
 from knowledge_table_api.models.graph import ExportData, Table
-from knowledge_table_api.services.llm import generate_schema
+from knowledge_table_api.services.llm_operations import generate_schema
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
