+++ backend/src/knowledge_table_api/services/llm/__init__.py
@@ -1,0 +1,1 @@
+"""Language model services."""
